Module rbj_eq::capi

A C interface to the filter library.

Structs

rbj_eq_coeffs_d_t

C API struct for filter coefficients.

rbj_eq_coeffs_df1_d_t

C API for filter coefficients in Direct Form I, suitable for running a filter.

rbj_eq_coeffs_df1_f_t

C API for filter coefficients in Direct Form I, suitable for running a filter.

rbj_eq_coeffs_f_t

C API struct for filter coefficients.

Enums

rbj_eq_t

C enum type for filters.

Functions

rbj_eq_coeffs_bandwidth_d

C API function to obtain filter coefficients for a filter specified by bandwidth value.

rbj_eq_coeffs_bandwidth_f

C API function to obtain filter coefficients for a filter specified by bandwidth value.

rbj_eq_coeffs_q_d

C API function to obtain filter coefficients for a filter specified by Q value.

rbj_eq_coeffs_q_f

C API function to obtain filter coefficients for a filter specified by Q value.

rbj_eq_coeffs_slope_d

C API function to obtain filter coefficients for a filter specified by slope value.

rbj_eq_coeffs_slope_f

C API function to obtain filter coefficients for a filter specified by slope value.

rbj_eq_coeffs_to_df1_d

C API for obtaining filter coefficients in Direct Form I from raw coefficients.

rbj_eq_coeffs_to_df1_f

C API for obtaining filter coefficients in Direct Form I from raw coefficients.

rbj_eq_df1_d

C API function to step a filter with the given Direct Form 1 coefficients and given state forward by one step given input x, returning the next y value.

rbj_eq_df1_f

C API function to step a filter with the given Direct Form 1 coefficients and given state forward by one step given input x, returning the next y value.