Crate rbj_eq

Background

Back in the day, DSP guru Robert Bristow-Johnson published a famous document titled Cookbook formulae for audio equalizer biquad filter coefficients. This is a really nice account of how to build “biquad” IIR filters useful for audio equalization, and also for a variety of other audio tasks. The RBJ filters are characterized by being extremely cheap to run, cheap to build on-the-fly, and having nice composability properties.

Many implementations of the RBJ filters exist in a variety of languages. This is the author’s implementation in Rust.

RBJ Filters

This crate provides implementations of the RBJ filters in safe Rust. What you get:

Function to compute filter coefficients for the various RBJ filter types.

Transfer function magnitude, derived from the coefficients.

A stateful filter function, based on the coefficients.

Examples

use rbj_eq::{LowPassFilter, FilterWidth};

// Make a sine wave at Nyquist.
let samples: Vec<f64> = (0..128)
    .map(|i| if i % 2 == 0 { 1.0 } else { -1.0 })
    .collect();

// Construct a half-band filter.
let cs = LowPassFilter.coeffs(
    0.5,
    FilterWidth::Slope {
        gain: 0.0,
        slope: 1.0,
    },
);
let mut filter = cs.make_filter();

// Filter the signal.
let filtered: Vec<f64> =
    samples.into_iter().map(|x| filter(x)).collect();

// The signal is damped. (The filter takes a few samples to converge.)
for (i, y) in filtered.iter().skip(4).enumerate() {
    assert!(y.abs() < 0.01, "filter fail: {i} {y}");
}

(See the examples directory of this distribution for more examples.)

Feature Flags

libm: Use the libm crate and its port of the MUSL floating point libraries to Rust, via the num-traits crate. This enables no_std.

Structs

FilterCoeffs

Biquad filter coefficients.

Enums

FilterWidth

Width / gain specification for filters.

Constants

AllPassFilter

All-pass filter.

BandNotchFilter

Bandnotch filter.

BandPassCFilter

Bandpass filter, with constant 0dB peak gain.

BandPassQFilter

Bandpass filter, with constant skirt gain and peak gain Q.

HighPassFilter

Highpass filter.

HighShelfFilter

Highpass shelf filter.

LowPassFilter

Lowpass filter.

LowShelfFilter

Lowpass shelf filter.

PeakingFilter

Peaking bandpass filter.